DON'T WORK - fullscreen start
Tried making the game fullscreen, doesn't work yet

+++ src/gamestate/GameStateManager.java
@@ -18,7 +18,7 @@ public class GameStateManager implements KeyListener, MouseWheelListener{
 	public static Screen s;
 	public GameStateManager(Run r){
 		this.r = r;
-		s = new Screen(r.frame.getWidth(),r.frame.getHeight());
+		s = new Screen((int) r.frame.getWidth(),r.frame.getHeight());
 		this.gs = new GameState(s,false);
 		moveScreen = new boolean[4];
 	}
